remove useless cast

+++ server/src/handler/eldenring/breakin.rs
@@ -40,7 +40,7 @@ impl HandleRequest<Box<RequestGetBreakInTargetListParams>, ResponseGetBreakInTar
         let mut pool_matches = self.services.pool_breakin.matches(&BreakInPoolQuery {
             player_id: self.session.player_id,
             play_region: request.play_region,
-            character_level: request.matching_parameters.character_level as u32,
+            character_level: request.matching_parameters.character_level,
             weapon_level: request.matching_parameters.max_reinforce as u32,
         });
